lots of readme
