--- .bumpversion.cfg
@@ -1,7 +1,0 @@
-[bumpversion]
-current_version = 0.1.0
-commit = True
-tag = False
-
-[bumpversion:file:pyproject.toml]
-[bumpversion:file:.github/workflows/docker.yaml]
