+++ pyproject.toml
@@ -22,13 +22,10 @@ classifiers=[
 
 dependencies=[
     'cpg-flow>=v0.2.4',
-    'gcsfs',
-    'fsspec',
     'hatchling>=1.27.0',
 ]
 
 [project.urls]
-# TODO(you): update this to reflect the project's official homepage
 Repository = "https://github.com/populationgenomics/cpg-flow-gatk-sv"
 
 [project.optional-dependencies]
@@ -42,7 +39,7 @@ test = [
 [project.scripts]
 
 # the workflow runner script - entrypoint for the pipeline
-run_workflow = 'cpg_flow_gatk_sv.first_workflow:cli_main'
+first_workflow = 'cpg_flow_gatk_sv.first_workflow:cli_main'
 
 [tool.setuptools]
 package-dir = {"" = "src"}
@@ -98,7 +95,6 @@ ignore = [
     "G004", # Logging statement uses f-string
     "PLW0603", # Using the global statement to update `<VAR>` is discouraged
     "Q000", # Single quotes found but double quotes preferred
-    "S101", # Use of assert detected
     "PLR0912", # Too many branches (> 12)
     "PLR0913", # Too many arguments in function (> 5)
     "C901", # method is too complex (> 10 conditions)
@@ -115,4 +111,28 @@ cpg = ["metamist", "cpg_flow", "cpg_utils", "cpg_flow_gatk_sv"]
 hail = ["hail"]
 
 [tool.ruff.lint.per-file-ignores]
-"src/cpg_flow_gatk_sv/first_stages.py" = ["ARG002"]
+"src/cpg_flow_gatk_sv/first_workflow.py" = ["ARG002"]
+
+[tool.bumpversion]
+current_version = "0.1.0"
+parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
+serialize = ["{major}.{minor}.{patch}"]
+commit = true
+message = "Bump version: {current_version} → {new_version}"
+commit_args = ""
+
+[[tool.bumpversion.files]]
+filename = "pyproject.toml"
+search = "version={current_version}"
+replace = "version={new_version}"
+
+[[tool.bumpversion.files]]
+filename = ".github/workflows/docker.yaml"
+search = "VERSION: {current_version}"
+replace = "VERSION: {new_version}"
+replace = "dragen_align_pa:{new_version}"
+
+[[tool.bumpversion.files]]
+filename = "README.md"
+search = "{current_version}"
+replace = "{new_version}"
