+++ src/cpg_flow_gatk_sv/config_template.toml
@@ -1,17 +1,68 @@
-# this is a template for the config file
-# demonstrating some of the common options used in cpg-flow
 [workflow]
+name = 'gatk_sv'
 
-# TODO(you): update this to reflect the workflow's name
-name = 'workflow_name'
-
-# mandatory fields for cpg-flow, not populated by analysis-runner
-# TODO(you): update this to contain a real cohort for the given sequencing type
-input_cohorts = []
-sequencing_type = 'genome'
-
-# used to make sure we don't repeat previously completed stages
 check_expected_outputs = true
 
-# the method to register outputs, can be missing - will not generate metamist analysis entries
+min_batch_size = 10
+
+sequencing_type = 'genome'
+
+# some genomes copied from Main which had succeeded through calling, to be used in test runs
+input_cohorts = ['COH5376']
+
+ref_fasta = 'gs://cpg-common-main/references/hg38/v0/dragen_reference/Homo_sapiens_assembly38_masked.fasta'
+
 status_reporter = 'metamist'
+time_limit_seconds = 86400  # 24 hrs max runtime
+
+# Use this option to run only a subset of jobs in the GatherSampleEvidence workflow
+# [workflow.GatherSampleEvidence]
+# only_jobs = ['coverage_counts', 'pesr', 'scramble', 'wham', 'manta']
+
+# these workflows take a VCF result from a GatherSampleEvidence stage, and optionally a baseline VCF
+# from the same caller, and runs a VCFMetrics comparison of the two
+# we don't have like-for-like baselines for any of our samples (AFAIK), and the workflows don't allow
+# us to recover these results
+[resource_overrides.GatherSampleEvidence]
+run_module_metrics = false
+
+[images]
+# cloned over from the GATK-SV config values
+cloud_sdk_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/cloud-sdk"
+cnmops_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/cnmops:2024-11-08-v1.0-62adb329"
+condense_counts_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/gatk:4.2.6.1-57-g9e03432"
+denovo = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/denovo:2025-03-25-v1.0.2-hotfix-65eddd78"
+gatk_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/gatk:2024-12-05-4.6.1.0-6-gfc248dfc1-NIGHTLY-SNAPSHOT"
+gatk_docker_pesr_override = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/gatk:4.2.6.1-57-g9e03432"
+genomes_in_the_cloud_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/genomes-in-the-cloud:2.3.2-1510681135"
+gq_recalibrator_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/gatk:mw-tb-form-sv-filter-training-data-899360a"
+linux_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/ubuntu1804"
+manta_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/manta:2023-09-14-v0.28.3-beta-3f22f94d"
+samtools_cloud_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/samtools-cloud:2024-10-25-v0.29-beta-5ea22a52"
+scramble_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/scramble:2024-10-25-v0.29-beta-5ea22a52"
+str = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/str:2023-05-23-v0.27.3-beta-e537bdd6"
+sv_base_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/sv-base:2024-10-25-v0.29-beta-5ea22a52"
+sv_base_mini_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/sv-base-mini:2024-10-25-v0.29-beta-5ea22a52"
+sv_pipeline_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/sv-pipeline:2025-03-25-v1.0.2-hotfix-65eddd78"
+sv_pipeline_qc_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/sv-pipeline:2025-03-25-v1.0.2-hotfix-65eddd78"
+wham_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/wham:2024-10-25-v0.29-beta-5ea22a52"
+
+# resource overrides for the GatherSampleEvidence stage
+[resource_overrides.GatherSampleEvidence.runtime_attr_scramble_part1]
+mem_gb = 7.5
+[resource_overrides.GatherSampleEvidence.runtime_attr_wham]
+mem_gb = 16
+[resource_overrides.GatherSampleEvidence.runtime_attr_scramble_part2]
+mem_gb = 16
+
+# resource overrides for the EvidenceQC stage
+[resource_overrides.EvidenceQC.wgd_build_runtime_attr]
+disk_gb = 40
+[resource_overrides.EvidenceQC.runtime_attr_mediancov]
+disk_gb = 200
+
+#[hail]
+#billing_project = "test-vcgs-dataset"
+#delete_scratch_on_exit = true
+#dry_run = true
+#backend = "local"
